Split Battle Cry analyzer into smaller modules
