+++ src/Parser/Warrior/Arms/CombatLogParser.js
@@ -3,7 +3,9 @@ import DamageDone from 'Parser/Core/Modules/DamageDone';
 
 import Abilities from './Modules/Abilities';
 import AlwaysBeCasting from './Modules/Features/AlwaysBeCasting';
-import BattleCry from './Modules/Features/BattleCry';
+import BattleCryDamage from './Modules/Core/BattleCryDamage';
+import BattleCryGCD from './Modules/Core/BattleCryGCD';
+import BattleCrySetup from './Modules/Core/BattleCrySetup';
 import ColossusSmash from './Modules/Core/ColossusSmash';
 import CooldownThroughputTracker from './Modules/Features/CooldownThroughputTracker';
 import ColossusSmashUptime from './Modules/BuffDebuff/ColossusSmashUptime';
@@ -27,7 +29,9 @@ class CombatLogParser extends CoreCombatLogParser {
 
     // Features
     alwaysBeCasting: AlwaysBeCasting,
-    battleCry: BattleCry,
+    battleCryDamage: BattleCryDamage,
+    battleCryGcd: BattleCryGCD,
+    battleCrySetup: BattleCrySetup,
     colossusSmash: ColossusSmash,
     cooldownThroughputTracker: CooldownThroughputTracker,
     colossusSmashUptime: ColossusSmashUptime,
